Serial port driver implemented
From now on, I do not have to put all debug messages on the vga monitor
because the messages can be sent through serial ports. In this commit,
logging of different device drivers is implemented. However, a better
disk reading routine is required to load larger kernel due to the size
has reached its maximum (25 kilobytes)

* Makefile: Serial output of qemu can be seen in terminal
* README.md: Update the work plan
* boot/bootsect.s: Increase the size of kernel to its maximum (25 KB)
* boot/setup.s: Likewise
* include/dev/serial.h: Created. Declaration of serial port drive can be
	found here
* kernel/Makefile: Refresh dependencies
* kernel/dev/serial.c: Created. Serial port driver's is implemented here
* kernel/dev/pic.c: Log information via serial ports
* kernel/dev/tty.c: Likewise
* kernel/kmalloc.c: Likewise
* kernel/main.c: Call `serial_init()' in order to use serial port driver